# api/v1/openapi.yaml
openapi: 3.0.3
info:
  title: Karabakh Atlas API
  description: >
    REST API serving region/city data for the Karabakh Atlas map website.
    Region/city write endpoints (POST/PUT/DELETE) require a logged-in
    session with the "admin" role (see /auth/login). Anyone can register
    an account via /auth/register, but new accounts get the "user" role,
    which can read everything but not write regions/cities.
  version: "1.0"
servers:
  - url: /api/v1
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
      description: >
        Session cookie set by POST /auth/login or /auth/register. The
        session user additionally needs the "admin" role for
        region/city write endpoints.
  schemas:
...
    City:
      type: object
      properties:
        id:
          type: string
          readOnly: true
        name:
          type: string
        region_id:
          type: string
        latitude:
          type: number
          format: float
        longitude:
          type: number
          format: float
        description:
          type: string
          nullable: true
        alt_names:
          type: string
          nullable: true
        image_url:
          type: string
          nullable: true
        image_credit:
          type: string
          nullable: true
        image_url_before:
          type: string
          nullable: true
          description: >
            An older/historical photo of the same place. When set
            alongside image_url, the frontend renders a before/after
            compare slider instead of a single photo.
        image_before_credit:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        website:
          type: string
          nullable: true
        category:
          type: string
          description: >
            "city" renders as a persistent map label; any other value is
            a user-added point of interest shown only when zoomed in.
          enum: [city, road, cafe, restaurant, hotel, landmark, museum,
                  park, university, school, hospital, pharmacy, bank,
                  government, police, fire_station, mosque, church,
                  fuel_station, parking, shop, grocery_store, other]
          default: city
        name_i18n:
          type: object
          nullable: true
          additionalProperties:
            type: string
        description_i18n:
          type: object
          nullable: true
          additionalProperties:
            type: string
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
    CityInput:
      type: object
      required: [name, latitude, longitude]
      properties:
        name:
          type: string
          maxLength: 128
        latitude:
          type: number
          format: float
          minimum: -90
          maximum: 90
        longitude:
          type: number
          format: float
          minimum: -180
          maximum: 180
        description:
          type: string
          nullable: true
        alt_names:
          type: string
          nullable: true
        image_url:
          type: string
          nullable: true
        image_credit:
          type: string
          nullable: true
        image_url_before:
          type: string
          nullable: true
        image_before_credit:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        website:
          type: string
          nullable: true
        category:
          type: string
          description: >
            "city" renders as a persistent map label; any other value is
            a user-added point of interest shown only when zoomed in.
            Defaults to "city" when omitted.
          enum: [city, road, cafe, restaurant, hotel, landmark, museum,
                  park, university, school, hospital, pharmacy, bank,
                  government, police, fire_station, mosque, church,
                  fuel_station, parking, shop, grocery_store, other]
        name_i18n:
          type: object
          nullable: true
          additionalProperties:
            type: string
        description_i18n:
          type: object
          nullable: true
          additionalProperties:
            type: string
...
paths:
...
  /regions/{region_id}/cities:
    parameters:
      - name: region_id
        in: path
        required: true
        schema:
          type: string
    get:
      summary: List cities in a region
      responses:
        "200":
          description: List of cities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/City"
        "404":
          description: Region not found
    post:
      summary: Create a city in a region
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CityInput"
      responses:
        "201":
          description: City created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/City"
        "400":
          description: Validation error
        "404":
          description: Region not found
  /cities:
    get:
      summary: List all cities
      parameters:
        - name: q
          in: query
          description: Filter by name (case-insensitive substring)
          schema:
            type: string
      responses:
        "200":
          description: List of cities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/City"
  /cities/{city_id}:
    parameters:
      - name: city_id
        in: path
        required: true
        schema:
          type: string
    get:
      summary: Get one city
      responses:
        "200":
          description: The city
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/City"
        "404":
          description: Not found
    put:
      summary: Update a city
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CityInput"
      responses:
        "200":
          description: Updated city
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/City"
        "400":
          description: Validation error
        "404":
          description: Not found
    delete:
      summary: Delete a city
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Deleted
        "404":
          description: Not found
...