# api/v1/views/cities.py
#!/usr/bin/python3
"""RESTful API views for City objects."""
import re
from api.v1.views import app_views
from flask import jsonify, abort, request
from models import storage
from models.region import Region
from models.city import City
from models.forum_post import ForumPost
from api.v1.forum_utils import delete_forum_posts
from models.favorite import Favorite
from models.image import Image
from api.v1.validation import (
    ValidationError,
    require_non_empty_string,
    require_number_in_range,
    optional_string,
    optional_url,
    optional_i18n_dict,
    optional_enum,
    only_allowed_fields,
)

# Matches the URL POST /images (api/v1/views/images.py) hands back, so
# an old photo can be told apart from a pasted external link — deleting
# or replacing one of *our* uploads should also drop the underlying
# Image row; a link to somewhere else isn't ours to delete.
_UPLOADED_IMAGE_ID_RE = re.compile(r"/api/v1/images/([^/?#]+)/?$")


def _delete_uploaded_image_if_ours(url):
    """Delete the Image row url points at, if it's one of our own
    uploads (no-op for an external link, a blank field, or an id that's
    already gone)."""
    if not url:
        return
    match = _UPLOADED_IMAGE_ID_RE.search(url)
    if not match:
        return
    image = storage.all(Image).get("Image.{}".format(match.group(1)))
    if image is not None:
        image.delete()


# "city" renders as a persistent map label; every other value is a
# user-added point of interest, rendered as a small marker only when
# zoomed in. Keep in sync with CITY_CATEGORIES in frontend/admin.js.
CITY_CATEGORIES = {
    "city", "road", "cafe", "restaurant", "hotel", "landmark", "museum",
    "park", "university", "school", "hospital", "pharmacy", "bank",
    "government", "police", "fire_station", "mosque", "church",
    "fuel_station", "parking", "shop", "grocery_store", "other",
}

CITY_FIELDS = {
    "name", "latitude", "longitude", "description", "alt_names",
    "image_url", "image_credit", "image_url_before", "image_before_credit",
    "phone", "website", "name_i18n", "description_i18n", "category",
}


def _validate_city_data(data, *, require_required_fields):
    """Validate a City request body; raises ValidationError on failure."""
    only_allowed_fields(data, CITY_FIELDS)
    if require_required_fields or "name" in data:
        require_non_empty_string(data, "name")
    if require_required_fields or "latitude" in data:
        require_number_in_range(data, "latitude", -90, 90)
    if require_required_fields or "longitude" in data:
        require_number_in_range(data, "longitude", -180, 180)
    optional_string(data, "description")
    optional_string(data, "alt_names", max_length=255)
    optional_url(data, "image_url", max_length=500)
    optional_string(data, "image_credit", max_length=255)
    optional_url(data, "image_url_before", max_length=500)
    optional_string(data, "image_before_credit", max_length=255)
    optional_string(data, "phone", max_length=30)
    optional_url(data, "website", max_length=500)
    optional_i18n_dict(data, "name_i18n")
    optional_i18n_dict(data, "description_i18n")
    # Unlike the optional_string/optional_url fields above, category
    # isn't something a client can legitimately clear — every City has
    # one — so an explicit null is rejected outright here rather than
    # falling through to optional_enum(), which treats None as "field
    # omitted" and would let it slip past validation only for the
    # unconditional setattr loop in update_city() below to actually set
    # city.category = None on the object.
    if "category" in data and data["category"] is None:
        raise ValidationError("category cannot be null")
    optional_enum(data, "category", CITY_CATEGORIES)


@app_views.route("/regions/<region_id>/cities", methods=["GET"])
def get_cities_by_region(region_id):
    """Return all City objects that belong to a given Region."""
    region = storage.all(Region).get("Region.{}".format(region_id))
    if region is None:
        abort(404)
    cities = [c for c in storage.all(City).values()
              if c.region_id == region_id]
    return jsonify([c.to_dict() for c in cities])


@app_views.route("/regions/<region_id>/cities", methods=["POST"])
def create_city(region_id):
    """Create a new City object under a given Region."""
    region = storage.all(Region).get("Region.{}".format(region_id))
    if region is None:
        abort(404)
    data = request.get_json(silent=True)
    if data is None:
        abort(400, description="Not a JSON")
    try:
        _validate_city_data(data, require_required_fields=True)
    except ValidationError as error:
        abort(400, description=error.message)
    data["region_id"] = region_id
    data.setdefault("category", "city")
    city = City(**data)
    city.save()
    return jsonify(city.to_dict()), 201


@app_views.route("/cities", methods=["GET"])
def get_cities():
    """Return all City objects, optionally filtered by name via ?q=."""
    search = request.args.get("q")
    cities = list(storage.all(City).values())
    if search:
        cities = [c for c in cities if search.lower() in c.name.lower()]
    return jsonify([c.to_dict() for c in cities])


@app_views.route("/cities/<city_id>", methods=["GET"])
def get_city(city_id):
    """Return a single City object by id, or 404 if not found."""
    city = storage.all(City).get("City.{}".format(city_id))
    if city is None:
        abort(404)
    return jsonify(city.to_dict())


@app_views.route("/cities/<city_id>", methods=["DELETE"])
def delete_city(city_id):
    """Delete a City object by id (and any forum posts or favorites
    referencing it), or 404 if not found."""
    city = storage.all(City).get("City.{}".format(city_id))
    if city is None:
        abort(404)
    delete_forum_posts([p for p in storage.all(ForumPost).values()
                        if p.target_city_id == city_id])
    for favorite in [f for f in storage.all(Favorite).values()
                     if f.city_id == city_id]:
        favorite.delete()
    _delete_uploaded_image_if_ours(city.image_url)
    _delete_uploaded_image_if_ours(city.image_url_before)
    city.delete()
    storage.save()
    return jsonify({}), 200


@app_views.route("/cities/<city_id>", methods=["PUT"])
def update_city(city_id):
    """Update an existing City object from a JSON body."""
    city = storage.all(City).get("City.{}".format(city_id))
    if city is None:
        abort(404)
    data = request.get_json(silent=True)
    if data is None:
        abort(400, description="Not a JSON")
    try:
        _validate_city_data(data, require_required_fields=False)
    except ValidationError as error:
        abort(400, description=error.message)
    image_url_fields = ("image_url", "image_url_before")
    for key, value in data.items():
        if key in image_url_fields and value != getattr(city, key):
            _delete_uploaded_image_if_ours(getattr(city, key))
        setattr(city, key, value)
    city.save()
    return jsonify(city.to_dict()), 200